+++ vue/src/index.ts
@@ -7,9 +7,7 @@
 
 import './onWidgetLoaded';
 
-export { default as ImportScheduler } from './ImportScheduler/ImportScheduler.vue';
 export { default as ImportSchedulerGA4 } from './ImportScheduler/ImportSchedulerGA4.vue';
-export { default as ImportSelector } from './ImportScheduler/ImportSelector.vue';
 export { default as ImportStatus } from './ImportStatus/ImportStatus.vue';
 export { default as ConfigureConnection } from './Configure/ConfigureConnection.vue';
 export { default as ClientConfig } from './ClientConfig/ClientConfig.vue';